pyproject.toml
[project]
name = "fxbot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "oandapyV20",
  "python-dotenv",
  "numpy",
  "scipy",
  "requests",
  "six",
]

[project.scripts]
fxbot = "fxbot.main:main"

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
